Implement deterministic daily rotation with queue
